Add files via upload
Ask a buyer to describe their dream home and they probably won't start with basement 
ceiling height or proximity to an east-west rail line. But the data set from this playground 
contest proves that price negotiations are influenced by much more than the number of 
bedrooms or a white picket fence.
This dataset contains 79 explanatory variables describing almost every aspect of residential 
homes in Ames, Iowa.
Goal: It is your job to predict the sales price for each house using everything you have 
learned so far. If you use a model not presented in class, you must justify it, explain how 
it works and describe precisely the role of each of the hyper-parameters. For each Id in 
the test set, you must predict the value of the SalePrice variable.